Add traverse_down_f32_max_mul and associates
This is the (R, max, *) semiring that is used in the Viterbi
algorithm. This can be used in dbentropy, where input is the
probability of traversing the edge from source to target. It is also
used in the knickpointfinder to compute maxima over reaches separated
by knickpoints by setting input equal to zero for any edge that
terminates on a knickpoint.

The max_mul_arg variants also keep track of the index of the maximum,
which is necessary for following the Viterbi path.