Define the NDEBUG macro in the Makefile (#201)
Resolves #200

This macro turns off the asserts that we use for testing in various
functions. CMake does this for us on many platforms, but R uses our
Makefile to compile libtopotoolbox, and for some reason NDEBUG does
not get properly defined on macos, causing the asserts to be left in
the compiled code, and R does not accept C code that has
asserts, because it could potentially abort the R session.

I have also added a test to the CI configuration that build the
Makefiles in CI on ubuntu and macos and check that no asserts are left
in the symbol table.